Changed Info.plist to say edu.berkeley.cnmat.matlabcommunicate  (for crash logs?) and made type signature use double instead of float for sample rate.

+++ externals/misc/matlabcommunicate/matlabcommunicate.c
@@ -558,7 +558,7 @@ done:
 }
 
 
-void matlabbridge_matlab2bufferchan(t_theobject *x, Symbol *matlabVariable, Symbol *buffername, int chan, float SR) {
+void matlabbridge_matlab2bufferchan(t_theobject *x, Symbol *matlabVariable, Symbol *buffername, int chan, double SR) {
   post("matlabcommunicate: matlab2bufferchan not implemented.");
 }
